Clean up commented codes

+++ src/Constants.java
@@ -2,7 +2,7 @@
  * Created by Xinan on 4/10/15.
  */
 public class Constants {
-  public final static short segmentSize = 1000;
-  public final static short headerSize = 10;
-  public final static short chunkSize = 990;
+  public final static int segmentSize = 1000;
+  public final static int headerSize = 10;
+  public final static int chunkSize = 990;
 }
